% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_read.R
\name{auto_read_tsv}
\alias{auto_read_tsv}
\title{auto_read_tsv()}
\usage{
auto_read_tsv(file, ...)
}
\arguments{
\item{file}{a file}

\item{...}{}
}
\value{
a tibble
}
\description{
auto_read_tsv()
}
\examples{
\dontrun{auto_read_tsv(file = "data-raw/depts2018.txt")}

}
