% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbles.R
\name{detect_na}
\alias{detect_na}
\title{Detect NA}
\usage{
detect_na(table)
}
\arguments{
\item{table}{name of the input variable}
}
\value{
a tibble with 3 columns (variable, n_missing, share_missing)
}
\description{
Detect NA takes a tibble and return the number and share of missing values for each variable
}
\examples{
dplyr::tbl(src = database_signauxfaibles, from = "table_training") \%>\%
dplyr::collect() \%>\%
detect_na()

}

