% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadetective.R
\name{detect_siren}
\alias{detect_siren}
\title{Detect Siren}
\usage{
detect_siren(.data)
}
\arguments{
\item{.data}{}
}
\value{
a tibble
}
\description{
detects if a data frame has a column of SIRENE identifier
}
\examples{

library("tibble")
table_datadetective <- tibble::tibble(v1 = c("123456789", "234567890"), v2 = c("1234567890123", "1234567890234"))
detect_siren(.data = table_datadetective)

}
