% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finder.R
\name{extract_all_functions}
\alias{extract_all_functions}
\title{Extract all functions}
\usage{
extract_all_functions(pkg)
}
\arguments{
\item{pkg}{name of a package}
}
\value{
a list of all functions in a package
}
\description{
Extract all functions
}
\examples{
extract_all_functions(pkg = "tricky")
extract_all_functions(pkg = "dplyr")
lsf.str("package:tricky")
ls("package:tricky")

}
