% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbles.R
\name{filter_dups_}
\alias{filter_dups_}
\title{Filter duplicates}
\usage{
filter_dups_(.data, .groups)
}
\arguments{
\item{.data}{a data frame}

\item{.groups}{a grouping variabme}
}
\value{
a data frame
}
\description{
filter_dups() takes a data frame and returns only duplicated rows
}
\examples{
table_test <- tibble::tibble(v1 = c("A", "A", "B", "C"), v2 = c("a", "b", "b", "c"))
filter_dups_(.data = table_test, .groups = ~ v1)
filter_dups_(.data = table_test, .groups = ~ v2)

}
