% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finder.R
\name{find_function_in_file}
\alias{find_function_in_file}
\title{Find a function in a file}
\usage{
find_function_in_file(function_name, file)
}
\arguments{
\item{function_name}{name of the function}

\item{file}{path to the file}
}
\value{
a tibble with two columns
}
\description{
Find a function in a file
}
\examples{
find_function_in_file(function_name = "detect_na", file = "R/tibbles.R")

}
