% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finder.R
\name{find_function_in_project}
\alias{find_function_in_project}
\title{Find function in the current RStudio project}
\usage{
find_function_in_project(function_name, files)
}
\arguments{
\item{function_name}{name of the project}

\item{files}{type of files : "R" or "Rmd"}
}
\value{
a tibble with two columns
}
\description{
Find function in the current RStudio project
}
\examples{
\dontrun{
find_function_in_project(function_name = "detect_na", file = "Rmd")
find_function_in_project(function_name = "detect_na", file = "R")
}

}
