% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{format_num}
\alias{format_num}
\title{Format num}
\usage{
format_num(x, locale, ...)
}
\arguments{
\item{x}{a number}

\item{locale}{a locale ("fr" or "en")}

\item{...}{}
}
\value{
a formatted number
}
\description{
Format num
}
\examples{
format_num(x = 1233.123, locale = "fr")
format_num(x = 1233.123, locale = "en")

}
