% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/french_formatting.R, R/numbers.R
\name{french_formatting}
\alias{french_formatting}
\alias{french_formatting}
\title{Formatting numbers according to French conventions}
\usage{
french_formatting(x)

french_formatting(x)
}
\arguments{
\item{x}{the number you need to format}

\item{x}{the number you need to format}
}
\value{
a string

a string
}
\description{
This function allows you to format numbers according to French conventions

This function allows you to format numbers according to French conventions
}
\examples{
french_formatting(11000.12)

french_formatting(11000.12)

}
\keyword{format}
