% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadetective.R
\name{is_siren}
\alias{is_siren}
\title{is SIREN}
\usage{
is_siren(var)
}
\arguments{
\item{var}{name of a variable}
}
\value{
a tibble with one boolean column
}
\description{
is_siren detect if the variable match the pattern of a SIRENE number
}
\examples{
library("tibble")
table_datadetective <- tibble(v1 = c("123456789", "234567890"), v2 = c("1234567890123", "1234567890234"))
is_siren(var = table_datadetective$v1)
is_siren(var = table_datadetective$v2)

}
