% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{make_single_node_node_table_}
\alias{make_single_node_node_table_}
\title{Make a single node node table}
\usage{
make_single_node_node_table_(.data, .id, node, link)
}
\arguments{
\item{.data}{a data frame with one row per (node, link)}

\item{.id}{id of a node}

\item{node}{name of the node variable}

\item{link}{name of the link variable}
}
\value{
a data frame with one row per node, node : ie one row per link
}
\description{
Make a single node node table
}
\examples{
## Toy data
toy_data <- tibble::tibble(
node = c("A", "B", "C", "A", "B"),
link = c("a", "a", "b", "b", "b")
)
make_single_node_node_table_(
.data = toy_data,
.id = "A",
node = "node",
link = "link"
)

}
