% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{parse_French_number}
\alias{parse_French_number}
\title{parse_French_numbers()}
\usage{
parse_French_number(x)
}
\arguments{
\item{x}{a string to be converted to a number}
}
\value{
a double
}
\description{
parse_French_numbers()
}
\examples{
parse_French_number(x = "1 123,123")

}
