% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_standardize}
\alias{str_standardize}
\title{Standardize}
\usage{
str_standardize(string)
}
\arguments{
\item{string}{a string}
}
\value{
a standardized string
}
\description{
Standardize removes accents, blank spaces, special caracters from a string
}
\examples{
library(readxl)
library(dplyr)
read_excel(path = "data-raw/DINSIC-Panorama_des_grands_projets_SI_20161116.xlsx") \%>\%
names() \%>\%
str_standardize(.))

}

