% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{studendize}
\alias{studendize}
\title{Studendize}
\usage{
studendize(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector
}
\description{
Normalizing a numerical vector using Student's normalization
}
\details{
WARNING : This version doesn't handles vectors with missing values
}
\examples{

studendize(x = rep(1,10))
studendize(x = 1:10) \%>\% mean()
studendize(x = 1:10) \%>\% sd()

}
