% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cartography.R
\name{theme_cartography}
\alias{theme_cartography}
\title{A simple theme for maps}
\usage{
theme_cartography(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{: size of the type}

\item{base_family}{: fonts family}
}
\description{
An elegant theme for maps with ggplot2
}
\examples{
ggplot() + theme_cartography()

}
\keyword{ggplot}
\keyword{maps}

