% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{week_of_the_year}
\alias{week_of_the_year}
\title{Compute the week of the year}
\usage{
week_of_the_year(date)
}
\arguments{
\item{date}{date}
}
\value{
a number
}
\description{
This function allows you to the week number in one year for a given date
}
\examples{
library("lubridate")
week_of_the_year(date = today())
week_of_the_year(date = as.Date("2015-10-21"))
week_of_the_year(seq(as.Date("1990-12-31"), as.Date("2015-12-31"), by = "year"))

}
\keyword{date}
\keyword{week}
