% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbles.R
\name{count_na}
\alias{count_na}
\title{Count NA}
\usage{
count_na(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a tibble
}
\description{
count_na count the number of NA in a vector
}
\examples{
read_excel(
path = system.file(
"extdata",
"DINSIC-Panorama_des_grands_projets_SI_20161116.xlsx",
package = "tricky")
) \%>\%
set_standard_names() \%>\%
.$ministere_porteur \%>\%
count_na()

}
