% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbles.R
\name{count_na}
\alias{count_na}
\title{Count NA}
\usage{
count_na(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a tibble
}
\description{
count_na count the number of NA in a vector
}
\examples{
plyr::ldply(.data = table_training, .fun = count_na)

}
