% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbles.R
\name{detect_na}
\alias{detect_na}
\title{Detect NA}
\usage{
detect_na(table)
}
\arguments{
\item{table}{name of the input variable}
}
\value{
a tibble with 3 columns (variable, n_missing, share_missing)
}
\description{
Detect NA takes a tibble and return the number and share of missing values for each variable
}
\examples{
read_excel(
path = system.file(
"extdata",
"DINSIC-Panorama_des_grands_projets_SI_20161116.xlsx",
package = "tricky")
) \%>\%
set_standard_names() \%>\%
detect_na()
}
