% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbles.R
\name{find_keys}
\alias{find_keys}
\title{Find keys}
\usage{
find_keys(table)
}
\arguments{
\item{table}{the name of the input table (either tibble or data.frame)}
}
\value{
a tibble with one column : keys
}
\description{
find_keys look at all function in a table and returns the list of possible keys (ie variables which identifies an observation)
}
\examples{
read_csv(system.file("extdata", "table_deputes.csv", package = "tricky")
)) \%>\% find_keys()

}
