% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/french_formatting.R
\name{french_formatting}
\alias{french_formatting}
\title{Formatting numbers according to French conventions}
\usage{
french_formatting(x)
}
\arguments{
\item{x}{the number you need to format}
}
\description{
This function allows you to format numbers according to French conventions
}
\examples{
french_formatting(11000.12)
}
\keyword{format}

