% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_formatting.R
\name{percent_formatting}
\alias{percent_formatting}
\title{Formatting percentages}
\usage{
percent_formatting(x)
}
\arguments{
\item{x}{the number you need to format}
}
\value{
a string
}
\description{
This function allows you to format percentages
}
\examples{
percent_formatting(0.012)


}
\keyword{format}
