% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbles.R
\name{set_standard_names}
\alias{set_standard_names}
\title{set standard names}
\usage{
set_standard_names(.data, prefix = "var_")
}
\arguments{
\item{.data}{a tibble}

\item{prefix}{name of the prefix for names beginning with numbers}
}
\value{
a tibble with standardized column names
}
\description{
set standard names
}
\examples{

library(readxl)
library(dplyr)
read_excel(
path = "data-raw/DINSIC-Panorama_des_grands_projets_SI_20161116.xlsx"
) \%>\%
set_standard_names()

}
