% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbles.R
\name{str_standardize}
\alias{str_standardize}
\title{Standardize table names}
\usage{
str_standardize(x)
}
\arguments{
\item{x}{a tibble}
}
\value{
a vector of standardized names
}
\description{
Names in tables may include blank caracters and many undesirable features.
str_standardize() makes it easy to standardize names and work with them inside R.
}
\examples{
library(readxl)
library(dplyr)
# source : http://www.data.gouv.fr/fr/datasets/panorama-des-grands-projets-si-de-letat-1/
read_excel(path = "raw-data/DINSIC-Panorama_des_grands_projets_SI_20161116.xlsx") \%>\%
setNames(str_standardize(.)) \%>\%
glimpse()

}
