% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_elegant.R
\name{theme_elegant}
\alias{theme_elegant}
\title{Elegant theme for ggplot2}
\usage{
theme_elegant(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{: size of the type}

\item{base_family}{: fonts family}
}
\description{
An elegant theme for ggplot2
}
\examples{
ggplot() + theme_elegant()
}
\keyword{ggplot}

