% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R, R/unfrench_formatting.R
\name{unfrench_formatting}
\alias{unfrench_formatting}
\alias{unfrench_formatting}
\title{Converts French formatted numbers into numbers}
\usage{
unfrench_formatting(x)

unfrench_formatting(x)
}
\arguments{
\item{a}{string containing a number}

\item{a}{string containing a number}
}
\value{
a number

a number
}
\description{
This function allows you to extract the value of French formatted numbers

This function allows you to extract the value of French formatted numbers
}
\examples{
unfrench_formatting(x = "0,12")

unfrench_formatting(x = "0,12")

}
\keyword{format}
